% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Experiment Simulation}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Iterator}

\item{n}{size}

\item{p}{probability}
}
\value{
simulation of bionomial experiment in barplots
}
\description{
Binomial Experiment Simulation
}
\examples{
iter <- 200
iter <- 500
}
