% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Poisson}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{Integer. Sample size.}

\item{iter}{Integer. Number of iterations.}

\item{lambda}{Numeric. Mean of the Poisson distribution.}

\item{...}{Additional arguments to be passed to \code{hist}.}
}
\value{
A plot with a histogram of sample means, a bar plot of sampled values, and a Poisson probability function.
}
\description{
Poisson
}
