% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{myf}
\alias{myf}
\title{Piecewise Linear Function}
\usage{
myf(x, coef)
}
\arguments{
\item{x}{numeric value or vector for independent variable}

\item{coef}{vector containing 3 values}
}
\value{
Numeric value or vector representing the output of the function
}
\description{
Piecewise Linear Function
}
\examples{
coef <- c(1, 2, 0.5)
}
