% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{display the curve, shaded area between the curve and x axis from -∞ to x=a}
\usage{
myncurve(mu=10, sigma=5, a=6)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{end point of x}
}
\value{
(probability, P(X<=a))
}
\description{
display the curve, shaded area between the curve and x axis from -∞ to x=a
}
