% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Overbook Simulation}
\usage{
ntickets(N= 200,gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{number of seats}

\item{gamma}{gamma}

\item{p}{Probability}
}
\value{
A list of results two plots
}
\description{
Overbook Simulation
}
